+++ core/adapters/containerengine/dockercompose/compositionRoot.go
@@ -7,7 +7,6 @@ import (
 )
 
 type compositionRoot interface {
-  InitOpUseCase() initOpUseCase
   RunOpUseCase() runOpUseCase
   KillOpRunUseCase() killOpRunUseCase
 }
@@ -16,7 +15,6 @@ func newCompositionRoot(
 ) (compositionRoot compositionRoot, err error) {
 
   filesys := _filesys{}
-  yamlCodec := _yamlCodec{}
 
   dockerEngine, err := dockerEngine.NewEnvClient()
   if (nil != err) {
@@ -28,7 +26,6 @@ func newCompositionRoot(
   opRunResourceFlusher := newOpRunResourceFlusher()
 
   compositionRoot = &_compositionRoot{
-    initOpUseCase: newInitOpUseCase(filesys, yamlCodec),
     runOpUseCase: newRunOpUseCase(opRunExitCodeReader, opRunResourceFlusher),
     killOpRunUseCase: newKillOpRunUseCase(opRunResourceFlusher, filesys),
   }
@@ -37,13 +34,8 @@ func newCompositionRoot(
 }
 
 type _compositionRoot struct {
-  initOpUseCase    initOpUseCase
   runOpUseCase     runOpUseCase
   killOpRunUseCase killOpRunUseCase
-}
-
-func (this _compositionRoot) InitOpUseCase() initOpUseCase {
-  return this.initOpUseCase
 }
 
 func (this _compositionRoot) RunOpUseCase() runOpUseCase {