+++ core/ports/containerEngine.go
@@ -5,24 +5,19 @@ package ports
 import "github.com/opctl/engine/core/logging"
 
 type ContainerEngine interface {
-  InitOp(
-  opBundlePath string,
-  name string,
-  ) (err error)
-
   RunOp(
   correlationId string,
   opArgs map[string]string,
   opBundlePath string,
   opName string,
-  opNamespace string,
+  opRunId string,
   logger logging.Logger,
-  ) (exitCode int, err error)
+  ) (err error)
 
   KillOpRun(
   correlationId string,
   opBundlePath string,
-  opNamespace string,
+  opRunId string,
   logger logging.Logger,
   ) (err error)
 }